--- docpac_15061224/partials/index.html
@@ -1,23 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Website</title>
-    <style>
-        * {
-            background-color: #181818;
-            color: whitesmoke;
-            font-family: Arial, sans-serif;
-            margin: 0;
-            padding: 0;
-        }
-    </style>
-</head>
-
-    <body>
-        Welcome to home page
-    </body>
-
-    </html>
