--- docpac_15061224/partials/views/index.ejs
@@ -1,7 +1,0 @@
-<%- include('partials/header') %>
-
-    <body>
-        Welcome to home page
-    </body>
-
-    <%- include('partials/footer') %>
